harjoitukset_ja_tehtavat_pythonilla/tehtava_3/requirements.txt
mysql-connector-python
flask
python-dotenv
psycopg2-binary
requests